Reorganise + add template
