+++ README.md
@@ -2,3 +2,7 @@ Xaitsnue Community Dictionary Project
 ========
 
 This website houses a dictionary of the Xaitsnue language created by members of the Xaitsnue community and [the UC Berkeley Department of Linguistics](https://lx.berkeley.edu/). It is based on [this template](https://eddersko.github.io/web-template/) by [@eddersko](https://github.com/eddersko).
+
+### Website
+
+To run locally / test, run `npm install` and then `npm run dev`. It should load up at  `localhost:3000`. Refer to the [Next.js documentation](https://nextjs.org/docs) for more, we're using the pages directory at the moment.
