+++ website/styles/globals.css
@@ -130,85 +130,19 @@
  body {
    font-family: HK Grotesk;
    background-color: var(--color-bg);
+   margin: 24px 16px;
  }
  
- /* Page structure */
- .wrapper {
-   min-height: var(--wrapper-height);
-   display: grid;
-   place-items: center;
-   margin: 0 1rem;
- }
- .content {
-   display: flex;
-   flex-direction: column;
-   align-items: center;
-   justify-content: center;
- }
- 
- /* Very light scaling for our illustration */
- .title {
-   color: #2800FF;
-   font-family: HK Grotesk;
-   font-style: normal;
-   font-weight: bold;
-   font-size: 100px;
-   line-height: 105%;
-   margin: 0;
- }
- 
- /* Very light scaling for our illustration */
- .illustration {
-   max-width: 100%;
-   max-height: var(--image-max-width);
-   margin-top: var(--image-margin);
- }
- 
- /* Instructions */
- .instructions {
-   margin: 1rem auto 0;
- }
- 
- /* Button - Add it from the README instructions */
  button,
  input {
-   font-family: inherit;
-   font-size: 100%;
-   background: #FFFFFF;
-   border: 1px solid #000000;
-   box-sizing: border-box;
-   border-radius: 4px;
-   padding: 0.5rem 1rem;
-   transition: 500ms;
- }
- 
- /* Subheading */
- h2 {
-   color: #2800FF;
- }
- 
- /* Interactive image */
- .illustration:active {
-   transform: translateY(5px);
- }
- 
- /* Button dip
- - Toggling this class on and off will move it down and up again
- - The button transition property above determines the speed of the translate (500ms)
- */
- .dipped {
-   transform: translateY(5px);
- }
- 
- /* Links that open files in the editor */
- .fileopener {
-   cursor:pointer;
-   font-weight:bold;
-   border-bottom:3px solid var(--color-primary);
-   color: var(--color-secondary);
- }
- .fileopener:hover {
-   border-bottom:3px solid var(--color-secondary);
+    font-family: inherit;
+    font-size: 100%;
+    background: #FFFFFF;
+    border: 1px solid #000000;
+    box-sizing: border-box;
+    border-radius: 4px;
+    padding: 0.5rem 1rem;
+    transition: 500ms;
  }
  
  input {
@@ -242,4 +176,26 @@
    background: rgba(166, 51, 214, 0.8);
    color: white;
    cursor: pointer;
+ }
+ 
+ main {
+   background: white;
+   margin: 16px 0px;
+   padding: 16px;
+   border-radius: 8px;
+ }
+ 
+ .nav {
+   display: flex;
+   gap: 16px;
+   font-size: 24px;
+ }
+ 
+ .nav > div:first-child {
+   flex-grow: 1;
+   font-weight: 800;
+ }
+ 
+ .nav > div > a {
+   text-decoration: none;
  }
